arreglada llamada y imagen

+++ resources/templates/contenido/AJAXRutinaFav.php
@@ -1,25 +1,20 @@
 <?php 
- 
-  print_r($_GET);
+ if(isset($_GET)){
   $id_user= clear_input($_GET['id_user']);
   $id_rutina= clear_input($_GET['id_rutina']);
   $fav= clear_input($_GET['fav']);
 
   if($fav != 'null'){
-    echo $fav;
     RutinaFavoritaManager::delete($fav);
-    print_r('quitado');
-    
   }else{
     RutinaFavoritaManager::insert(intval($id_rutina),intval($id_user));
-    print_r('insertado');
-
-    
   }
-
-
-
-
-
-
+  $res = RutinaFavoritaManager::getByIdRutina($id_rutina,$id_user);
+  if($res == null){
+    $resultado = 'null';
+  }else{
+    $resultado = $res['ID'];
+  }
+  print_r($resultado);
+}
 ?>
